Eliminate the concept of value_weight.
Speeds up list operations substantially, and maps somewhat.

Ironically, this whole system was intended to make lists and maps faster, by
allowing them to accurately estimate how much memory they were wasting. This
was a flawed concept since doing so would require an O(n) scan of the wasted
values, at which point we might as well just do a non-persistent copy. The
other option is to average the weight, but at that point the weight itself
becomes moot.

The value caching system was also part of the rationale, but that plan has
changed as well.