Upgrade to LLVM 3.7.
That was painful, but the pain points definitely weren't where I
expected.

One nicity is that the newer version seems to have assertions enabled in
the FreeBSD package (or there's just more enabled since *I* have them
enabled and far more of LLVM is inlined in headers now), so a number of
problems were uncovered that way. (Mostly surrounding LLVM blindly
casting scopes to DISubprograms.)