Permit isolated control macros to run in void contexts.
This was originally going to dovetail into treating isolated Blocks as
inline code, which would be useful for macros. However, that would
result in people shooting themselves in the foot on the accasion where
they use such a macro alone within a lambda, which would result in the
intended inline block itself becomming a lambda. (Other less subtle
misuses would also result in the creation of lambdas.)

Since this is a desireable feature, keep it anyway.